backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "This repository hosts the Agentic Framework, designed to accelerate the development of quick agent solutions from scratch."
authors = ["Julian Boronat <julian.boronat@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
tqdm = "^4.67.1"
pandas = "^2.2.3"
langchain-aws = "^0.2.15"
langchain-mongodb = "^0.5.0"
langgraph = "^0.3.5"
grandalf = "^0.8"
langgraph-checkpoint-mongodb = "^0.1.1"
annotated-types = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
